+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add support for arithmetic operators
 * Add support for accessing properties using bracket notation
+* Update `CONTAINS` operator to support objects
 
 ## 1.2.0 (2019-04-11)